requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0
albumentations=1.3.1
appnope=0.1.3
asttokens=2.2.1
attrs=23.1.0
backcall=0.2.0
backports=1.0
backports.functools_lru_cache=1.6.5
blas=1.0
blinker=1.6.2
blis=0.7.9
brotlipy=0.7.0
ca-certificates=2023.05.30
cachetools=5.3.1
catalogue=2.0.8
certifi=2022.12.7
cffi=1.15.1
chardet=4.0.0
charset-normalizer=3.1.0
click=8.1.3
colorama=0.4.6
comet-ml=3.33.5
confection=0.1.0
configobj=5.0.8
contourpy=1.1.0
cryptography=39.0.1
cycler=0.10.0
cymem=2.0.7
debugpy=1.5.1
decorator=5.1.1
dulwich=0.21.5
easyocr=1.7.0
entrypoints=0.4
everett=3.1.0
executing=1.2.0
fastai=2.7.12
fastcore=1.5.29
fastdownload=0.0.7
fastprogress=1.0.3
filelock=3.12.2
flask=2.3.2
fonttools=4.40.0
freetype=2.12.1
giflib=5.2.1
gmp=6.2.1
gmpy2=2.1.2
google-auth=2.22.0
google-auth-oauthlib=1.0.0
grpcio=1.56.2
gunicorn=21.2.0
icecream=2.1.3
idna=2.10
imageio=2.31.1
imgaug=0.4.0
importlib-metadata=6.8.0
importlib-resources=5.12.0
ipykernel=6.15.0
ipynb=0.5.1
ipython=8.12.0
itsdangerous=2.1.2
jedi=0.18.2
jinja2=3.1.2
joblib=1.3.1
jsonschema=4.17.3
jupyter_client=7.3.4
jupyter_core=5.3.0
kiwisolver=1.4.4
langcodes=3.3.0
lazy-loader=0.3
lcms2=2.12
lerc=3.0
libcxx=14.0.6
libdeflate=1.17
libffi=3.4.4
libgfortran=5.0.0
libgfortran5=11.3.0
libopenblas=0.3.21
libpng=1.6.39
libsodium=1.0.18
libtiff=4.5.0
libwebp=1.2.4
libwebp-base=1.2.4
llvm-openmp=14.0.6
lz4-c=1.9.4
markdown=3.4.4
markdown-it-py=3.0.0
markupsafe=2.1.3
matplotlib=3.7.1
matplotlib-inline=0.1.6
mdurl=0.1.2
mpc=1.1.0
mpfr=4.0.2
mpmath=1.3.0
murmurhash=1.0.9
ncurses=6.4
nest-asyncio=1.5.6
networkx=3.1
ninja=1.11.1
numpy=1.24.4
numpy-base=1.24.3
oauthlib=3.2.2
opencv-python=4.8.0.74
opencv-python-headless=4.8.0.74
openssl=3.1.1
packaging=23.1
pandas=2.0.3
parso=0.8.3
pathy=0.10.2
pexpect=4.8.0
pickleshare=0.7.5
pillow=10.0.0
pillow-heif=0.13.0
pip=23.2.1
pkgutil-resolve-name=1.3.10
platformdirs=3.8.0
preshed=3.0.8
prompt-toolkit=3.0.38
prompt_toolkit=3.0.38
protobuf=4.23.4
psutil=5.9.0
ptyprocess=0.7.0
pure_eval=0.2.2
py-cpuinfo=9.0.0
pyasn1=0.5.0
pyasn1-modules=0.3.0
pyclipper=1.3.0.post4
pycparser=2.21
pydantic=1.10.10
pygments=2.15.1
pyheif=0.7.1
pyopenssl=23.0.0
pyparsing=2.4.7
pyrsistent=0.19.3
pysocks=1.7.1
python=3.8.16
python-bidi=0.4.2
python-box=6.1.0
python-dateutil=2.8.2
python-dotenv=1.0.0
python_abi=3.8
pytz=2023.3
pywavelets=1.4.1
pyyaml=6.0
pyzmq=25.1.0
qudida=0.0.4
readline=8.2
requests=2.31.0
requests-oauthlib=1.3.1
requests-toolbelt=1.0.0
rich=13.4.2
roboflow=1.1.0
rsa=4.9
scikit-image=0.21.0
scikit-learn=1.3.0
scipy=1.10.1
seaborn=0.12.2
semantic-version=2.10.0
sentry-sdk=1.26.0
setuptools=67.8.0
shapely=2.0.1
simplejson=3.19.1
six=1.16.0
smart-open=6.3.0
spacy=3.5.4
spacy-legacy=3.0.12
spacy-loggers=1.0.4
sqlite=3.41.2
srsly=2.4.6
stack_data=0.6.2
supervision=0.11.1
sympy=1.12
tensorboard=2.13.0
tensorboard-data-server=0.7.1
thinc=8.1.10
threadpoolctl=3.1.0
tifffile=2023.4.12
tk=8.6.12
torch=2.0.1
torchaudio=0.11.0
torchvision=0.12.0
tornado=6.1
tqdm=4.65.0
traitlets=5.9.0
typer=0.9.0
typing-extensions=4.7.1
typing_extensions=4.7.1
tzdata=2023.3
ultralytics=8.0.165
urllib3=2.0.3
wasabi=1.1.2
wcwidth=0.2.6
websocket-client=1.3.3
werkzeug=2.3.6
wget=3.2
wheel=0.38.4
wrapt=1.15.0
wurlitzer=3.0.3
xz=5.4.2
zeromq=4.3.4
zipp=3.15.0
zlib=1.2.13
zstd=1.5.5